instrumentation-genai/opentelemetry-instrumentation-langchain/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "opentelemetry-instrumentation-langchain"
dynamic = ["version"]
description = "OpenTelemetry Official Langchain instrumentation"
readme = "README.rst"
license = "Apache-2.0"
requires-python = ">=3.8"
dependencies = [
  "opentelemetry-api ~= 1.30",
  "opentelemetry-instrumentation ~= 0.51b0",
  "opentelemetry-semantic-conventions ~= 0.51b0",
]

[project.optional-dependencies]
instruments = ["langchain>=0.0.346", "langchain-core>0.1.0"]
test = [
  "pytest",
  "pytest-sugar",
  "vcrpy",
  "pytest-recording",
  "pytest-asyncio",
  "langchain>=0.0.346",
  "langchain-core>0.1.0",
  "langchain-openai",
  "langchain-community",
]

[project.entry-points.opentelemetry_instrumentor]
langchain = "opentelemetry.instrumentation.langchain:LangchainInstrumentor"

[tool.hatch.version]
path = "src/opentelemetry/instrumentation/langchain/version.py"

[tool.hatch.build.targets.sdist]
include = ["/src", "/tests"]

[tool.hatch.build.targets.wheel]
packages = ["src/opentelemetry"] 